+++ resources/views/comments/view.blade.php
@@ -1,9 +1,9 @@
 <div class="pt-4">
-    <hr/>
+    <hr class="border-gray-200 dark:border-white/10"/>
 
     <div class="fi-fo-field-wrp py-8">
         <div class="grid gap-y-4">
-        <span class="text-sm font-medium leading-6 text-gray-950 dark:text-white">
+        <span class="text-base font-semibold leading-6 text-gray-950 dark:text-white">
             {{ __('finisterre::finisterre.comments.title') }}
         </span>